Move images to respective directories
